requirements.txt
# Core Dependencies
gradio==4.19.2
pandas==2.2.0
numpy==1.26.3

# File Processing Libraries
PyMuPDF==1.23.21
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text Processing & NLP
nltk==3.8.1
regex==2023.12.25

# Utilities
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
chardet==5.2.0


pip3 install PyMuPDF python-docx openpyxl PyPDF2 pdfplumber nltk regex python-magic chardet